[jsk_fetch_startup] parameterize workspace path and make it as argument

+++ jsk_fetch_robot/jsk_fetch_startup/scripts/update_workspace.sh
@@ -1,21 +1,46 @@
 #!/usr/bin/env bash
 
-. $HOME/ros/melodic/devel/setup.bash
+function usage()
+{
+    echo "Usage: $0 [-w workspace_directory] [-h]"
+}
+
+WORKSPACE=$HOME/ros/melodic/
+
+while getopts hw: OPT
+do
+    case $OPT in
+        w)
+            WORKSPACE=$(cd $(dirname $OPTARG) && pwd)/$(basename $OPTARG)
+            ;;
+        h)
+            usage
+            exit 1
+            ;;
+    esac
+done
+
+if [ ! -e $WORKSPACE ]; then
+    echo "No workspace $WORKSPACE"
+    exit 1
+fi
+
+. $WORKSPACE/devel/setup.bash
 # Filename should end with .txt to preview the contents in a web browser
-LOGFILE=$HOME/ros/melodic/update_workspace.txt
+LOGFILE=$WORKSPACE/update_workspace.txt
 
 {
 set -x
 # Update workspace
-cd $HOME/ros/melodic/src
-ln -sf $(rospack find jsk_fetch_startup)/../jsk_fetch.rosinstall.$ROS_DISTRO $HOME/ros/melodic/src/.rosinstall
+cd $WORKSPACE/src
+ln -sf $(rospack find jsk_fetch_startup)/../jsk_fetch.rosinstall.$ROS_DISTRO $WORKSPACE/src/.rosinstall
 wstool foreach --git 'git stash'
 wstool update --delete-changed-uris
 wstool foreach --git 'branch-name=$(git rev-parse --abbrev-ref HEAD) && git reset --hard HEAD && git checkout origin/$branch-name && git branch -D $branch-name && git checkout $branch-name' # Forcefully checkout specified branch
 wstool update
 WSTOOL_UPDATE_RESULT=$?
 # Build workspace
-cd $HOME/ros/melodic
+cd $WORKSPACE
 catkin clean aques_talk collada_urdf_jsk_patch libcmt -y
 catkin init
 catkin config -DCMAKE_BUILD_TYPE=Release
